improve config models and remove allowed expects from telemetry
