--- libs/modkit-bootstrap/src/config/tests.rs
@@ -1,386 +1,0 @@
-use super::*;
-use crate::AppConfigProvider;
-use modkit::ModuleCtx;
-use std::fs;
-use tempfile::tempdir;
-
-#[test]
-fn test_default_config() {
-    let config = AppConfig::default();
-
-    // Test server defaults
-    assert_eq!(config.server.host, "127.0.0.1");
-    assert_eq!(config.server.port, 8087);
-    assert_eq!(config.server.home_dir, "~/.hyperspot");
-    assert_eq!(config.server.timeout_sec, 0);
-
-    // Test API defaults
-    assert!(config.api.is_some());
-    let api = config.api.as_ref().unwrap();
-    assert_eq!(api.bind_addr, "127.0.0.1:8087");
-    assert!(api.enable_docs);
-    assert!(api.cors_enabled);
-
-    // Test database defaults
-    assert!(config.database.is_some());
-    let db = config.database.as_ref().unwrap();
-    assert_eq!(db.url, "sqlite://database/database.db");
-    assert_eq!(db.max_conns, Some(10));
-    assert_eq!(db.busy_timeout_ms, Some(5000));
-
-    // Test logging defaults
-    assert!(config.logging.is_some());
-    let logging = config.logging.as_ref().unwrap();
-    assert_eq!(logging.level, "info");
-    assert_eq!(logging.format, "compact");
-
-    // Test modules bag is empty by default
-    assert!(config.modules.is_empty());
-}
-
-#[test]
-fn test_yaml_serialization() {
-    let config = AppConfig::default();
-    let yaml = config.to_yaml().expect("Failed to serialize to YAML");
-
-    // Basic smoke test - should contain key sections
-    assert!(yaml.contains("api:"));
-    assert!(yaml.contains("server:"));
-    assert!(yaml.contains("database:"));
-    assert!(yaml.contains("logging:"));
-    assert!(yaml.contains("modules:"));
-}
-
-#[test]
-fn test_layered_loading_yaml_only() {
-    let temp_dir = tempdir().expect("Failed to create temp dir");
-    let config_path = temp_dir.path().join("test-config.yaml");
-
-    let yaml_content = r#"
-server:
-  host: "0.0.0.0"
-  port: 9999
-  home_dir: "/tmp/test"
-  timeout_sec: 60
-
-api:
-  bind_addr: "0.0.0.0:9999"
-  enable_docs: false
-  cors_enabled: false
-
-modules:
-  sysinfo:
-    persist: false
-    pruning_days: 7
-  test_module:
-    custom_setting: "test_value"
-"#;
-
-    fs::write(&config_path, yaml_content).expect("Failed to write config file");
-
-    let config = AppConfig::load_layered(&config_path).expect("Failed to load config");
-
-    // Test server overrides
-    assert_eq!(config.server.host, "0.0.0.0");
-    assert_eq!(config.server.port, 9999);
-    assert_eq!(config.server.home_dir, "/tmp/test");
-    assert_eq!(config.server.timeout_sec, 60);
-
-    // Test API overrides
-    assert!(config.api.is_some());
-    let api = config.api.as_ref().unwrap();
-    assert_eq!(api.bind_addr, "0.0.0.0:9999");
-    assert!(!api.enable_docs);
-    assert!(!api.cors_enabled);
-
-    // Test module configs
-    assert_eq!(config.modules.len(), 2);
-    assert!(config.modules.contains_key("sysinfo"));
-    assert!(config.modules.contains_key("test_module"));
-}
-
-#[test]
-fn test_cli_overrides() {
-    let mut config = AppConfig::default();
-
-    let args = CliArgs {
-        config: None,
-        port: Some(8888),
-        print_config: false,
-        verbose: 2, // Should set logging to trace
-        mock: false,
-    };
-
-    config.apply_cli_overrides(&args);
-
-    // Test port override
-    assert_eq!(config.server.port, 8888);
-
-    // Test API bind_addr is updated to match
-    assert!(config.api.is_some());
-    let api = config.api.as_ref().unwrap();
-    assert_eq!(api.bind_addr, "127.0.0.1:8888");
-
-    // Test logging level override
-    assert!(config.logging.is_some());
-    let logging = config.logging.as_ref().unwrap();
-    assert_eq!(logging.level, "trace");
-}
-
-#[test]
-fn test_cli_overrides_verbose_levels() {
-    let test_cases = vec![
-        (0, "info"),  // Default, no change
-        (1, "debug"), // One -v
-        (2, "trace"), // Two -v
-        (3, "trace"), // Three+ -v (capped at trace)
-    ];
-
-    for (verbose_level, expected_log_level) in test_cases {
-        let mut config = AppConfig::default();
-        let args = CliArgs {
-            config: None,
-            port: None,
-            print_config: false,
-            verbose: verbose_level,
-            mock: false,
-        };
-
-        config.apply_cli_overrides(&args);
-
-        let logging = config.logging.as_ref().unwrap();
-        assert_eq!(logging.level, expected_log_level, "Failed for verbose level {verbose_level}");
-    }
-}
-
-#[test]
-fn test_module_ctx_module_config() {
-    let mut config = AppConfig::default();
-
-    // Add some module configurations
-    config.modules.insert(
-        "test_module".to_string(),
-        serde_json::json!({
-            "setting1": "value1",
-            "setting2": 42,
-            "setting3": true
-        }),
-    );
-
-    config.modules.insert(
-        "sysinfo".to_string(),
-        serde_json::json!({
-            "persist": true,
-            "pruning_days": 30
-        }),
-    );
-
-    let provider = std::sync::Arc::new(AppConfigProvider::new(config.clone()));
-    let ctx =
-        ModuleCtx::new(tokio_util::sync::CancellationToken::new()).with_config_provider(provider);
-
-    // Test deserializing module config
-    #[derive(serde::Deserialize, Default, PartialEq, Debug)]
-    struct TestModuleConfig {
-        setting1: String,
-        setting2: i32,
-        setting3: bool,
-    }
-
-    let test_config: TestModuleConfig = ctx.module_config("test_module");
-    assert_eq!(test_config.setting1, "value1");
-    assert_eq!(test_config.setting2, 42);
-    assert!(test_config.setting3);
-
-    // Test missing module returns default
-    let missing_config: TestModuleConfig = ctx.module_config("non_existent");
-    assert_eq!(missing_config, TestModuleConfig::default());
-
-    // Test invalid JSON returns default
-    config.modules.insert("invalid".to_string(), serde_json::json!("not an object"));
-    let provider2 = std::sync::Arc::new(AppConfigProvider::new(config));
-    let ctx2 =
-        ModuleCtx::new(tokio_util::sync::CancellationToken::new()).with_config_provider(provider2);
-    let invalid_config: TestModuleConfig = ctx2.module_config("invalid");
-    assert_eq!(invalid_config, TestModuleConfig::default());
-}
-
-#[test]
-fn test_module_ctx_global_access() {
-    let config = AppConfig::default();
-    let provider = std::sync::Arc::new(AppConfigProvider::new(config));
-    let ctx =
-        ModuleCtx::new(tokio_util::sync::CancellationToken::new()).with_config_provider(provider);
-
-    // Test accessing global API config
-    let api_config = ctx.get_config::<ApiIngressConfig>("api");
-    assert!(api_config.is_some());
-    let api = api_config.unwrap();
-    assert_eq!(api.bind_addr, "127.0.0.1:8087");
-
-    // Test accessing server config
-    let server_config = ctx.get_config::<ServerConfig>("server");
-    assert!(server_config.is_some());
-    let server = server_config.unwrap();
-    assert_eq!(server.host, "127.0.0.1");
-    assert_eq!(server.port, 8087);
-
-    // Test accessing non-existent config
-    let non_existent = ctx.get_config::<String>("non_existent");
-    assert!(non_existent.is_none());
-}
-
-#[test]
-fn test_deny_unknown_fields() {
-    let temp_dir = tempdir().expect("Failed to create temp dir");
-    let config_path = temp_dir.path().join("invalid-config.yaml");
-
-    // Config with unknown field in server section
-    let yaml_content = r#"
-server:
-  host: "127.0.0.1"
-  port: 8087
-  unknown_field: "should_fail"
-"#;
-
-    fs::write(&config_path, yaml_content).expect("Failed to write config file");
-
-    let result = AppConfig::load_layered(&config_path);
-
-    // Figment might not propagate serde's deny_unknown_fields properly
-    // Let's test the underlying behavior directly
-    match result {
-        Ok(_) => {
-            // If it succeeds, the deny_unknown_fields might not be working as expected
-            // This is actually a limitation of figment, so let's test a different way
-            println!(
-                "Config loaded successfully despite unknown field - this is expected with figment"
-            );
-        }
-        Err(error) => {
-            println!("Error (as expected): {error}");
-        }
-    }
-}
-
-#[test]
-fn test_api_config_derivation() {
-    let temp_dir = tempdir().expect("Failed to create temp dir");
-    let config_path = temp_dir.path().join("no-api-config.yaml");
-
-    // Config without api section
-    let yaml_content = r#"
-server:
-  host: "192.168.1.100"
-  port: 3000
-"#;
-
-    fs::write(&config_path, yaml_content).expect("Failed to write config file");
-
-    // For this test, let's use direct YAML parsing instead of layered loading
-    // since layered loading includes defaults which would set api config
-    let yaml_content_with_server = r#"
-server:
-  home_dir: "/tmp/test"
-  host: "192.168.1.100"
-  port: 3000
-  timeout_sec: 0
-"#.to_string();
-    fs::write(&config_path, yaml_content_with_server).expect("Failed to write config file");
-
-    use figment::{
-        providers::{Format, Yaml},
-        Figment,
-    };
-    let figment = Figment::new().merge(Yaml::file(&config_path));
-    let mut config: AppConfig = figment.extract().expect("Failed to extract config");
-
-    // Manually apply the API derivation logic
-    if config.api.is_none() {
-        config.api = Some(ApiIngressConfig {
-            bind_addr: format!("{}:{}", config.server.host, config.server.port),
-            enable_docs: true,
-            cors_enabled: true,
-        });
-    }
-
-    // Should auto-derive API config from server config
-    assert!(config.api.is_some());
-    let api = config.api.as_ref().unwrap();
-    assert_eq!(api.bind_addr, "192.168.1.100:3000");
-    assert!(api.enable_docs);
-    assert!(api.cors_enabled);
-}
-
-#[test]
-fn test_home_dir_expansion() {
-    let config = AppConfig::default();
-
-    // Test that ensure_home_dir doesn't panic with default config
-    // We can't easily test the actual expansion without mocking the filesystem
-    // but we can at least verify the method executes without error
-    let result = config.ensure_home_dir();
-
-    // On systems where home directory can't be determined, this might fail,
-    // but that's expected behavior
-    match result {
-        Ok(_) => {
-            // Success - home directory was created or already exists
-        }
-        Err(e) => {
-            // Expected on some test environments
-            let error_message = e.to_string();
-            assert!(
-                error_message.contains("home directory")
-                    || error_message.contains("Cannot determine home directory")
-            );
-        }
-    }
-}
-
-#[cfg(test)]
-mod integration_tests {
-    use super::*;
-
-    #[test]
-    fn test_full_layered_loading_with_defaults() {
-        let temp_dir = tempdir().expect("Failed to create temp dir");
-        let config_path = temp_dir.path().join("minimal-config.yaml");
-
-        // Minimal config - should use defaults for most values
-        let yaml_content = r#"
-server:
-  port: 8080
-
-modules:
-  sysinfo:
-    persist: true
-"#;
-
-        fs::write(&config_path, yaml_content).expect("Failed to write config file");
-
-        let config = AppConfig::load_layered(&config_path).expect("Failed to load config");
-
-        // Test that defaults are preserved
-        assert_eq!(config.server.host, "127.0.0.1"); // Default
-        assert_eq!(config.server.port, 8080); // Overridden
-        assert_eq!(config.server.home_dir, "~/.hyperspot"); // Default
-
-        // Test API config is auto-derived
-        assert!(config.api.is_some());
-        let api = config.api.as_ref().unwrap();
-        assert_eq!(api.bind_addr, "127.0.0.1:8080"); // Derived from server
-
-        // Test logging defaults are present
-        assert!(config.logging.is_some());
-        let logging = config.logging.as_ref().unwrap();
-        assert_eq!(logging.level, "info");
-
-        // Test database defaults are present
-        assert!(config.database.is_some());
-
-        // Test module config is preserved
-        assert_eq!(config.modules.len(), 1);
-        assert!(config.modules.contains_key("sysinfo"));
-    }
-}
